+++ src/internal/decl/hash-decl.h
@@ -1,0 +1,1 @@
+static r_obj* hash_file_impl(void* p_data);