+++ src/internal/file.h
@@ -1,0 +1,8 @@
+#ifndef RLANG_INTERNAL_FILE_H
+#define RLANG_INTERNAL_FILE_H
+
+#include <stdio.h>
+
+FILE* r_fopen(r_obj* path, const char* mode);
+
+#endif