+++ src/internal/hash.c
@@ -1,4 +1,5 @@
 #include <rlang.h>
+#include "file.h"
 
 /*
  * Using the standard xxhash defines, as seen in:
@@ -11,6 +12,8 @@
 
 #include <stdio.h> // sprintf()
 #include <inttypes.h> // PRIx64
+
+#include "decl/hash-decl.h"
 
 /*
  * Construct a define specifying whether version 2 or 3 of
@@ -88,6 +91,7 @@ struct hash_state_t {
 
 static inline struct hash_state_t new_hash_state(XXH3_state_t* p_xx_state);
 static inline int hash_version();
+static inline r_obj* hash_value(XXH3_state_t* p_xx_state);
 static inline void hash_bytes(R_outpstream_t stream, void* p_input, int n);
 static inline void hash_char(R_outpstream_t stream, int input);
 
@@ -130,18 +134,11 @@ r_obj* hash_impl(void* p_data) {
 
   R_Serialize(x, &stream);
 
-  XXH128_hash_t hash = XXH3_128bits_digest(p_xx_state);
+  r_obj* value = KEEP(hash_value(p_xx_state));
+  r_obj* out = r_str_as_character(value);
 
-  // R assumes C99, so these are always defined as `uint64_t` in xxhash.h
-  XXH64_hash_t high = hash.high64;
-  XXH64_hash_t low = hash.low64;
-
-  // 32 for hash, 1 for terminating null added by `sprintf()`
-  char out[32 + 1];
-
-  sprintf(out, "%016" PRIx64 "%016" PRIx64, high, low);
-
-  return r_chr(out);
+  FREE(1);
+  return out;
 }
 
 static
@@ -170,6 +167,22 @@ int hash_version() {
 #else
   return 2;
 #endif
+}
+
+static inline
+r_obj* hash_value(XXH3_state_t* p_xx_state) {
+  XXH128_hash_t hash = XXH3_128bits_digest(p_xx_state);
+
+  // R assumes C99, so these are always defined as `uint64_t` in xxhash.h
+  XXH64_hash_t high = hash.high64;
+  XXH64_hash_t low = hash.low64;
+
+  // 32 for hash, 1 for terminating null added by `sprintf()`
+  char out[32 + 1];
+
+  sprintf(out, "%016" PRIx64 "%016" PRIx64, high, low);
+
+  return r_str(out);
 }
 
 static inline void hash_skip(struct hash_state_t* p_state, void* p_input, int n);
@@ -246,3 +259,149 @@ void hash_skip(struct hash_state_t* p_state, void* p_input, int n) {
 #endif // USE_VERSION_3
 
 #undef USE_VERSION_3
+
+// -----------------------------------------------------------------------------
+
+r_obj* ffi_hash_file(r_obj* path) {
+  XXH3_state_t* p_xx_state = XXH3_createState();
+
+  struct exec_data data = {
+    .x = path,
+    .p_xx_state = p_xx_state
+  };
+
+  return R_ExecWithCleanup(hash_file_impl, &data, hash_cleanup, &data);
+}
+
+#define CHUNK_SIZE 512 * 1024
+
+static
+r_obj* hash_file_impl(void* p_data) {
+  struct exec_data* p_exec_data = (struct exec_data*) p_data;
+  r_obj* path = p_exec_data->x;
+  XXH3_state_t* p_xx_state = p_exec_data->p_xx_state;
+
+  if (r_typeof(path) != R_TYPE_character) {
+    r_abort("`path` must be a character vector.");
+  }
+
+  r_ssize n_path = r_length(path);
+  r_obj* const* v_path = r_chr_cbegin(path);
+
+  r_obj* out = KEEP(r_alloc_character(n_path));
+
+  // Allocate before opening file to avoid handle leak on allocation failure
+  void* buf = (void*)R_alloc(CHUNK_SIZE, sizeof(char));
+
+  for (r_ssize i = 0; i < n_path; ++i) {
+    XXH_errorcode err = XXH3_128bits_reset(p_xx_state);
+    if (err == XXH_ERROR) {
+      r_abort("Can't initialize hash state.");
+    }
+
+    r_obj* elt = v_path[i];
+
+    FILE* fp = r_fopen(elt, "rb");
+    if (fp == NULL) {
+      r_abort("Can't open file: %s.", Rf_translateChar(elt));
+    }
+
+    size_t n_read;
+
+    while ((n_read = fread(buf, 1, CHUNK_SIZE, fp)) > 0) {
+      XXH_errorcode err = XXH3_128bits_update(p_xx_state, buf, n_read);
+
+      if (err == XXH_ERROR) {
+        fclose(fp);
+        r_abort("Can't update hash state.");
+      }
+    }
+
+    fclose(fp);
+
+    r_chr_poke(out, i, hash_value(p_xx_state));
+  }
+
+  FREE(1);
+  return out;
+}
+
+#undef CHUNK_SIZE
+
+// -----------------------------------------------------------------------------
+
+static inline
+void hasher_finalizer(r_obj* x) {
+  void* p_x = R_ExternalPtrAddr(x);
+
+  if (!p_x) {
+    // Defensively exit if the external pointer resolves to `NULL`
+    return;
+  }
+
+  XXH3_state_t* p_xx_state = (XXH3_state_t*) p_x;
+  XXH3_freeState(p_xx_state);
+
+  R_ClearExternalPtr(x);
+}
+
+r_obj* ffi_hasher_init() {
+  XXH3_state_t* p_xx_state = XXH3_createState();
+
+  XXH_errorcode err = XXH3_128bits_reset(p_xx_state);
+  if (err == XXH_ERROR) {
+    r_abort("Can't initialize hash state.");
+  }
+
+  SEXP out = KEEP(R_MakeExternalPtr(p_xx_state, r_null, r_null));
+
+  R_RegisterCFinalizerEx(out, hasher_finalizer, TRUE);
+
+  FREE(1);
+  return(out);
+}
+
+r_obj* ffi_hasher_update(r_obj* x, r_obj* data) {
+  if (r_typeof(x) != R_TYPE_pointer) {
+    r_abort("`x` must be a hasher.");
+  }
+
+  if (r_typeof(data) != R_TYPE_raw) {
+    r_abort("`data` must be a raw vector.");
+  }
+
+  void* p_x = R_ExternalPtrAddr(x);
+
+  if (!p_x) {
+    r_abort("`x` must be a hasher.");
+  }
+
+  XXH3_state_t* p_xx_state = (XXH3_state_t*) p_x;
+
+  void* v_data = (void*) r_raw_begin(data);
+  int size = r_ssize_as_integer(r_length(data));
+
+  XXH_errorcode err = XXH3_128bits_update(p_xx_state, v_data, size);
+
+  if (err == XXH_ERROR) {
+    r_abort("Can't update hash state.");
+  }
+
+  return r_true;
+}
+
+r_obj* ffi_hasher_value(r_obj* x) {
+  if (r_typeof(x) != R_TYPE_pointer) {
+    r_abort("`x` must be a hasher.");
+  }
+
+  void* p_x = R_ExternalPtrAddr(x);
+
+  if (!p_x) {
+    r_abort("`x` must be a hasher.");
+  }
+
+  XXH3_state_t* p_xx_state = (XXH3_state_t*) p_x;
+
+  return hash_value(p_xx_state);
+}